backend/requirements/base.txt
asgiref==3.8.1
celery==5.5.3
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyparsing==3.2.3
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2